+++ plugins/pixart-rf/fu-pxi-common.c
@@ -129,9 +129,9 @@ gchar *
 fu_pxi_hpac_version_info_parse(const guint16 hpac_ver)
 {
 	return g_strdup_printf("%u%u.%u%u.%u",
-			       hpac_ver / 10000,
-			       (hpac_ver / 1000) % 10,
-			       ((hpac_ver / 100) % 10),
-			       (hpac_ver / 10) % 10,
-			       hpac_ver % 10);
+			       (guint)(hpac_ver / 10000),
+			       (guint)((hpac_ver / 1000) % 10),
+			       (guint)((hpac_ver / 100) % 10),
+			       (guint)((hpac_ver / 10) % 10),
+			       (guint)(hpac_ver % 10));
 }
